+++ mido/messages/__init__.py
@@ -1,4 +1,6 @@
 from .specs import SPEC_LOOKUP, SPEC_BY_TYPE, SPEC_BY_STATUS
+from .specs import MIN_PITCHWHEEL, MAX_PITCHWHEEL, MIN_SONGPOS, MAX_SONGPOS
 from .checks import check_time
 from .messages import BaseMessage, Message
 from .messages import parse_string, format_as_string, parse_string_stream
+